services/quantification/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
requests==2.31.0
pandas
